Generate a single code action with relative path
