+++ src/service/getGenerateSnippetCodeAction.ts
@@ -20,58 +20,59 @@ import { stepDefinitionSnippet } from './snippet/stepDefinitionSnippet.js'
  *
  * @param diagnostics all the diagnostics
  * @param link where the snippet should be added
+ * @param relativePath the relative path from the workspace root
  * @param createFile true if link.targetUri does not exist
  * @param mustacheTemplate template to generae the snippet
  * @param languageName the name of the language we're generating for
  * @param registry parameter types
  */
-export function getGenerateSnippetCodeActions(
+export function getGenerateSnippetCodeAction(
   diagnostics: Diagnostic[],
   link: LocationLink,
+  relativePath: string,
   createFile: boolean,
   mustacheTemplate: string | undefined,
   languageName: LanguageName,
   registry: ParameterTypeRegistry
-): CodeAction[] {
+): CodeAction | null {
   const undefinedStepDiagnostic = diagnostics.find((d) => d.code === diagnosticCodeUndefinedStep)
   const language = getLanguage(languageName)
   const stepKeyword = undefinedStepDiagnostic?.data?.stepKeyword
   const stepText = undefinedStepDiagnostic?.data?.stepText
-  if (undefinedStepDiagnostic && stepText) {
-    const generator = new CucumberExpressionGenerator(() => registry.parameterTypes)
-    const generatedExpressions = generator.generateExpressions(stepText)
-
-    const snippet = stepDefinitionSnippet(
-      stepKeyword,
-      generatedExpressions,
-      mustacheTemplate || language.defaultSnippetTemplate,
-      language.snippetParameters
-    )
-
-    const documentChanges: (CreateFile | TextDocumentEdit)[] = []
-    if (createFile) {
-      documentChanges.push(
-        CreateFile.create(link.targetUri, {
-          ignoreIfExists: true,
-          overwrite: true,
-        })
-      )
-    }
-    documentChanges.push(
-      TextDocumentEdit.create(VersionedTextDocumentIdentifier.create(link.targetUri, 0), [
-        TextEdit.replace(link.targetRange, snippet),
-      ])
-    )
-    const ca: CodeAction = {
-      title: 'Generate step definition',
-      diagnostics: [undefinedStepDiagnostic],
-      kind: CodeActionKind.QuickFix,
-      edit: {
-        documentChanges,
-      },
-      isPreferred: true,
-    }
-    return [ca]
+  if (!undefinedStepDiagnostic || !stepText) {
+    return null
   }
-  return []
+  const generator = new CucumberExpressionGenerator(() => registry.parameterTypes)
+  const generatedExpressions = generator.generateExpressions(stepText)
+
+  const snippet = stepDefinitionSnippet(
+    stepKeyword,
+    generatedExpressions,
+    mustacheTemplate || language.defaultSnippetTemplate,
+    language.snippetParameters
+  )
+
+  const documentChanges: (CreateFile | TextDocumentEdit)[] = []
+  if (createFile) {
+    documentChanges.push(
+      CreateFile.create(link.targetUri, {
+        ignoreIfExists: true,
+        overwrite: true,
+      })
+    )
+  }
+  documentChanges.push(
+    TextDocumentEdit.create(VersionedTextDocumentIdentifier.create(link.targetUri, 0), [
+      TextEdit.replace(link.targetRange, snippet),
+    ])
+  )
+  return {
+    title: `Define in ${relativePath}`,
+    diagnostics: [undefinedStepDiagnostic],
+    kind: CodeActionKind.QuickFix,
+    edit: {
+      documentChanges,
+    },
+    isPreferred: true,
+  }
 }
